+++ src/components/common/Modals/CommentModal/types.ts
@@ -1,9 +1,9 @@
-import type { Post } from '../../../../types/domain/models'
+import type { PostWithInteractions } from '../../../../types/domain/models'
 
 export type CommentModalProps = {
   isOpen: boolean
   onClose: () => void
-  originalPost: Post
+  originalPost: PostWithInteractions
   userName: string
   userAvatar: string | null
 }