+++ src/store/slices/api/posts/index.ts
@@ -1,5 +1,4 @@
 export * from './posts.api'
-export * from './comments.api'
 export * from './likes.api'
 export * from './retweets.api'
 export * from './replies.api'